tests/lib: Resolve circular import
Commit 769b54c8e76126f6 introduced a circular import. Resolve this by
moving the hardware and node base classes into extra files.
