+++ tests/lib/hardware/hardware_base.py
@@ -27,7 +27,6 @@
 from abc import ABC, abstractmethod
 import os
 import tempfile
-from typing import Dict, Any
 import uuid
 
 import paramiko.rsakey
@@ -95,38 +94,3 @@ class HardwareBase(ABC):
             self._ansible_runner_nodes = self.nodes.copy()
 
         return self.ansible_runner.run_play(play_source)
-
-
-class NodeBase(ABC):
-    """
-    Base class for nodes
-    """
-    def __init__(self, name: str, private_key: str):
-        self.name = name
-        self.private_key = private_key
-
-    @abstractmethod
-    def get_ssh_ip(self) -> str:
-        """
-        Get the IP address that can be used to ssh into the node
-        """
-        pass
-
-    def ansible_inventory_vars(self) -> Dict[str, Any]:
-        vars = {
-            'ansible_host': self.get_ssh_ip(),
-            # FIXME(jhesketh): Set username depending on OS
-            'ansible_user': config.NODE_IMAGE_USER,
-            'ansible_ssh_private_key_file': self.private_key,
-            'ansible_host_key_checking': False,
-            'ansible_ssh_host_key_checking': False,
-            'ansible_scp_extra_args': '-o StrictHostKeyChecking=no',
-            'ansible_ssh_extra_args': '-o StrictHostKeyChecking=no',
-            'ansible_python_interpreter': '/usr/bin/python3',
-            'ansible_become': False,
-        }
-        if config.NODE_IMAGE_USER != "root":
-            vars['ansible_become'] = True
-            vars['ansible_become_method'] = 'sudo'
-            vars['ansible_become_user'] = 'root'
-        return vars
